+++ demo/admin/src/config.ts
@@ -1,4 +1,6 @@
-import cometConfig from "../comet-config.json";
+import { SiteConfig } from "@comet/cms-admin";
+
+import cometConfig from "./comet-config.json";
 import environment from "./environment";
 
 export function createConfig() {
@@ -18,8 +20,10 @@ export function createConfig() {
         ...cometConfig,
         apiUrl: environmentVariables.API_URL,
         adminUrl: environmentVariables.ADMIN_URL,
-        sitesConfig: JSON.parse(environmentVariables.SITES_CONFIG),
+        sitesConfig: JSON.parse(environmentVariables.SITES_CONFIG) as SitesConfig,
     };
 }
+
+export type SitesConfig = Record<string, SiteConfig>;
 
 export type Config = ReturnType<typeof createConfig>;