+++ packages/admin/cms-admin/src/common/MasterMenu.tsx
@@ -1,28 +1,46 @@
-import { Menu, MenuCollapsibleItem, MenuContext, MenuItemRouterLink, MenuItemRouterLinkProps, useWindowSize } from "@comet/admin";
+import {
+    Menu,
+    MenuCollapsibleItem,
+    MenuContext,
+    MenuItemAnchorLink,
+    MenuItemAnchorLinkProps,
+    MenuItemRouterLink,
+    MenuItemRouterLinkProps,
+    useWindowSize,
+} from "@comet/admin";
 import * as React from "react";
 import { RouteProps, useRouteMatch } from "react-router-dom";
 
-import { CurrentUserContext } from "../userPermissions/hooks/currentUser";
+import { useUserPermissionCheck } from "../userPermissions/hooks/currentUser";
 
-export type MasterMenuItem = Omit<MenuItemRouterLinkProps, "to"> & {
+type MasterMenuItemRoute = Omit<MenuItemRouterLinkProps, "to"> & {
     requiredPermission?: string;
     route?: RouteProps;
     to?: string;
     submenu?: MasterMenuItem[];
 };
 
+type MasterMenuItemAnchor = MenuItemAnchorLinkProps & {
+    requiredPermission?: string;
+};
+
+export type MasterMenuItem = MasterMenuItemRoute | MasterMenuItemAnchor;
+
 export type MasterMenuData = MasterMenuItem[];
 
+export function isMasterMenuItemAnchor(item: MasterMenuItem): item is MasterMenuItemAnchor {
+    return !!item.href;
+}
+
 export function useMenuFromMasterMenuData(items: MasterMenuData): MenuItem[] {
-    const context = React.useContext(CurrentUserContext);
-    const checkPermission = (item: MasterMenuItem): boolean => {
-        if (!item.requiredPermission) return true;
-        if (context === undefined)
-            throw new Error("MasterMenu: requiredPermission is set but CurrentUserContext not found. Make sure CurrentUserProvider exists.");
-        return context.isAllowed(context.currentUser, item.requiredPermission);
-    };
+    const isAllowed = useUserPermissionCheck();
+    const checkPermission = (item: MasterMenuItem): boolean => !item.requiredPermission || isAllowed(item.requiredPermission);
 
     const mapFn = (item: MasterMenuItem): MenuItem => {
+        if (isMasterMenuItemAnchor(item)) {
+            return { menuItem: item };
+        }
+
         const { route, submenu, to, ...menuItem } = item;
         return {
             menuItem: {
@@ -36,11 +54,21 @@ export function useMenuFromMasterMenuData(items: MasterMenuData): MenuItem[] {
     return items.filter(checkPermission).map(mapFn);
 }
 
-type MenuItem = {
+type MenuItemRoute = {
     menuItem: MenuItemRouterLinkProps;
     hasSubmenu: boolean;
     submenu: MenuItem[];
 };
+
+type MenuItemAnchor = {
+    menuItem: MenuItemAnchorLinkProps;
+};
+
+type MenuItem = MenuItemRoute | MenuItemAnchor;
+
+function isMenuItemAnchor(item: MenuItem): item is MenuItemAnchor {
+    return !!item.menuItem.href;
+}
 
 export interface MasterMenuProps {
     permanentMenuMinWidth?: number;
@@ -66,11 +94,17 @@ export const MasterMenu: React.FC<MasterMenuProps> = ({ menu, permanentMenuMinWi
     return (
         <Menu variant={useTemporaryMenu ? "temporary" : "permanent"}>
             {menuItems.map((menuItem, index) =>
-                menuItem.hasSubmenu ? (
+                isMenuItemAnchor(menuItem) ? (
+                    <MenuItemAnchorLink key={index} {...menuItem.menuItem} />
+                ) : menuItem.hasSubmenu ? (
                     <MenuCollapsibleItem key={index} {...menuItem.menuItem}>
-                        {menuItem.submenu.map((submenu, index) => (
-                            <MenuItemRouterLink key={index} {...submenu.menuItem} to={`${match.url}${submenu.menuItem.to}`} />
-                        ))}
+                        {menuItem.submenu.map((submenuItem, index) =>
+                            isMenuItemAnchor(submenuItem) ? (
+                                <MenuItemAnchorLink key={index} {...submenuItem.menuItem} />
+                            ) : (
+                                <MenuItemRouterLink key={index} {...submenuItem.menuItem} to={`${match.url}${submenuItem.menuItem.to}`} />
+                            ),
+                        )}
                     </MenuCollapsibleItem>
                 ) : (
                     <MenuItemRouterLink key={index} {...menuItem.menuItem} to={`${match.url}${menuItem.menuItem.to}`} />