+++ packages/admin/cms-admin/src/dam/DataGrid/fileUpload/useDamFileUpload.tsx
@@ -9,6 +9,7 @@ import { NetworkError, UnknownError } from "../../../common/errors/errorMessages
 import { upload } from "../../../form/file/upload";
 import { useDamAcceptedMimeTypes } from "../../config/useDamAcceptedMimeTypes";
 import { useDamScope } from "../../config/useDamScope";
+import { clearDamItemCache } from "../../helpers/clearDamItemCache";
 import { FilenameData, useManualDuplicatedFilenamesHandler } from "../duplicatedFilenames/ManualDuplicatedFilenamesHandler";
 import { NewlyUploadedItem, useFileUploadContext } from "./FileUploadContext";
 import { FileUploadErrorDialog } from "./FileUploadErrorDialog";
@@ -21,25 +22,24 @@ import {
     UnsupportedTypeError,
 } from "./fileUploadErrorMessages";
 import { ProgressDialog } from "./ProgressDialog";
-import { createDamFolderForFolderUpload, damFolderByNameAndParentId } from "./useFileUpload.gql";
+import { createDamFolderForFolderUpload, damFolderByNameAndParentId } from "./useDamFileUpload.gql";
 import {
     GQLDamFolderByNameAndParentIdQuery,
     GQLDamFolderByNameAndParentIdQueryVariables,
     GQLDamFolderForFolderUploadMutation,
     GQLDamFolderForFolderUploadMutationVariables,
-} from "./useFileUpload.gql.generated";
+} from "./useDamFileUpload.gql.generated";
 
 interface FileWithPath extends File {
     path?: string;
 }
 
-interface FileWithFolderPath extends FileWithPath {
+export interface FileWithFolderPath extends FileWithPath {
     folderPath?: string;
 }
 
-interface UploadFileOptions {
+interface UploadDamFileOptions {
     acceptedMimetypes?: string[];
-    onAfterUpload?: (errorOccurred: boolean) => void;
 }
 
 interface Files {
@@ -47,8 +47,18 @@ interface Files {
     fileRejections: FileRejection[];
 }
 
+type ImportSource = { importSourceType: never; importSourceId: never } | { importSourceType: string; importSourceId: string };
+
+interface UploadFilesOptions {
+    folderId?: string;
+    importSource?: ImportSource;
+}
+
 export interface FileUploadApi {
-    uploadFiles: ({ acceptedFiles, fileRejections }: Files, folderId?: string) => void;
+    uploadFiles: (
+        { acceptedFiles, fileRejections }: Files,
+        { folderId, importSource }: UploadFilesOptions,
+    ) => Promise<{ hasError: boolean; rejectedFiles: RejectedFile[]; uploadedItems: NewlyUploadedItem[] }>;
     validationErrors?: FileUploadValidationError[];
     maxFileSizeInBytes: number;
     dialogs: React.ReactNode;
@@ -63,6 +73,10 @@ export interface FileUploadApi {
 export interface FileUploadValidationError {
     file: Pick<FileWithFolderPath, "name" | "path">;
     message: React.ReactNode;
+}
+
+interface RejectedFile {
+    file: File;
 }
 
 const addFolderPathToFiles = async (acceptedFiles: FileWithPath[]): Promise<FileWithFolderPath[]> => {
@@ -98,7 +112,7 @@ const addFolderPathToFiles = async (acceptedFiles: FileWithPath[]): Promise<File
     return newFiles;
 };
 
-export const useFileUpload = (options: UploadFileOptions): FileUploadApi => {
+export const useDamFileUpload = (options: UploadDamFileOptions): FileUploadApi => {
     const context = useCmsBlockContext(); // TODO create separate CmsContext?
     const client = useApolloClient();
     const manualDuplicatedFilenamesHandler = useManualDuplicatedFilenamesHandler();
@@ -351,17 +365,22 @@ export const useFileUpload = (options: UploadFileOptions): FileUploadApi => {
         [manualDuplicatedFilenamesHandler],
     );
 
-    const uploadFiles = async ({ acceptedFiles, fileRejections }: Files, folderId?: string) => {
+    const uploadFiles = async (
+        { acceptedFiles, fileRejections }: Files,
+        { folderId, importSource }: UploadFilesOptions,
+    ): Promise<{ hasError: boolean; rejectedFiles: RejectedFile[]; uploadedItems: NewlyUploadedItem[] }> => {
         setProgressDialogOpen(true);
         setValidationErrors(undefined);
 
         const uploadedFolders: Array<NewlyUploadedItem & { type: "folder" }> = [];
         const uploadedFiles: Array<NewlyUploadedItem & { type: "file" }> = [];
+        const rejectedFiles: Array<RejectedFile> = [];
 
         let errorOccurred = false;
         if (fileRejections.length > 0) {
             errorOccurred = true;
             generateValidationErrorsForRejectedFiles(fileRejections);
+            rejectedFiles.push(...fileRejections.map(({ errors, ...rejection }) => rejection));
         }
 
         if (acceptedFiles.length > 0) {
@@ -397,6 +416,8 @@ export const useFileUpload = (options: UploadFileOptions): FileUploadApi => {
                             file,
                             folderId: targetFolderId,
                             scope,
+                            importSourceId: importSource?.importSourceId,
+                            importSourceType: importSource?.importSourceType,
                         },
                         cancelUpload.current.token,
                         {
@@ -406,7 +427,7 @@ export const useFileUpload = (options: UploadFileOptions): FileUploadApi => {
                         },
                     );
 
-                    uploadedFiles.push({ id: response.data.id, parentId: targetFolderId, type: "file" });
+                    uploadedFiles.push({ id: response.data.id, parentId: targetFolderId, type: "file", file });
                 } catch (err) {
                     errorOccurred = true;
                     const typedErr = err as AxiosError<{ error: string; message: string; statusCode: number }>;
@@ -433,8 +454,13 @@ export const useFileUpload = (options: UploadFileOptions): FileUploadApi => {
                     } else {
                         addValidationError(file, <UnknownError />);
                     }
+
+                    rejectedFiles.push({ file });
                 }
             }
+
+            await client.reFetchObservableQueries();
+            clearDamItemCache(client.cache);
         }
 
         setProgressDialogOpen(false);
@@ -443,9 +469,10 @@ export const useFileUpload = (options: UploadFileOptions): FileUploadApi => {
         }
         setTotalSizes({});
         setUploadedSizes({});
-        options.onAfterUpload?.(errorOccurred);
 
         addNewlyUploadedItems([...uploadedFolders, ...uploadedFiles]);
+
+        return { hasError: errorOccurred, rejectedFiles: rejectedFiles, uploadedItems: [...uploadedFolders, ...uploadedFiles] };
     };
 
     return {